pyproject.toml
[project]
name = "flub-agent"
version = "0.1.0"
description = "Multi-agent travel planning system with Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk",
    "python-dotenv",
    "anyio",
    "mcp",
    "fast-flights>=2.2",
    "primp",
    "protobuf",
    "selectolax",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
